--- frontend/src/diary/components/data_section.jsx
@@ -1,6 +1,0 @@
-function DataSection(){
-    return (
-        <div className="w-2/3 h-auto ml-4 mr-4 bg-white rounded-lg">데이터섹션</div>
-    );
-}
-export default DataSection;
